[feat/#1178] 구글로그인 시점 변경
